backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
requests==2.31.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
schedule==1.2.1
playwright==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0